Moved filter implementation to abstract hydrator
