+++ library/Zend/Stdlib/Hydrator/ClassMethods.php
@@ -11,7 +11,6 @@
 namespace Zend\Stdlib\Hydrator;
 
 use Zend\Stdlib\Exception;
-use Zend\Stdlib\Hydrator\Filter\FilterComposite;
 use Zend\Stdlib\Hydrator\Filter\GetFilter;
 use Zend\Stdlib\Hydrator\Filter\HasFilter;
 use Zend\Stdlib\Hydrator\Filter\IsFilter;
@@ -30,12 +29,6 @@ class ClassMethods extends AbstractHydrator implements HydratorOptionsInterface
     protected $underscoreSeparatedKeys = true;
 
     /**
-     * Composite to filter the methods, that need to be hydrated
-     * @var Filter\FilterComposite
-     */
-    protected $filterComposite;
-
-    /**
      * Define if extract values will use camel case or name with underscore
      * @param boolean|array $underscoreSeparatedKeys
      */
@@ -44,7 +37,6 @@ class ClassMethods extends AbstractHydrator implements HydratorOptionsInterface
         parent::__construct();
         $this->setUnderscoreSeparatedKeys($underscoreSeparatedKeys);
 
-        $this->filterComposite = new FilterComposite();
         $this->filterComposite->addFilter("is", new IsFilter());
         $this->filterComposite->addFilter("has", new HasFilter());
         $this->filterComposite->addFilter("get", new GetFilter());
@@ -174,54 +166,4 @@ class ClassMethods extends AbstractHydrator implements HydratorOptionsInterface
         return $object;
     }
 
-    /**
-     * Add a new filter to take care of what needs to be hydrated.
-     * To exclude e.g. the method getServiceLocator:
-     *
-     * <code>
-     * $composite->addFilter("servicelocator",
-     *     function($property) {
-     *         list($class, $method) = explode('::', $property);
-     *         if ($method === 'getServiceLocator') {
-     *             return false;
-     *         }
-     *         return true;
-     *     }, FilterComposite::CONDITION_AND
-     * );
-     * </code>
-     *
-     * @param string $name Index in the composite
-     * @param callable|Zend\Stdlib\Hydrator\Filter\FilterInterface $filter
-     * @param int $condition
-     */
-    public function addFilter($name, $filter, $condition = FilterComposite::CONDITION_OR)
-    {
-        $this->filterComposite->addFilter($name, $filter, $condition);
-    }
-
-    /**
-     * Check whether a specific filter exists at key $name or not
-     *
-     * @param string $name Index in the composite
-     * @return bool
-     */
-    public function hasFilter($name)
-    {
-        return $this->filterComposite->hasFilter($name);
-    }
-
-    /**
-     * Remove a filter from the composition.
-     * To not extract "has" methods, you simply need to unregister it
-     *
-     * <code>
-     * $filterComposite->removeFilter('has');
-     * </code>
-     *
-     * @param $name
-     */
-    public function removeFilter($name)
-    {
-        $this->filterComposite->removeFilter($name);
-    }
 }
